Adding TouchTabStrip
This patch adds a basic TouchTabStrip for touch_ui builds.  It is intended to be a placeholder for further work towards a more refined TouchTabStrip.  It is modeled after the existing TabStrip and tries to make use of BaseTabStrip and BaseTab where appropriate.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/5572007
Patch from Chad Faragher <wyck@chromium.org>.