Fix uninitialized variables introduced in CL 72374.
time_slice_hrs is being used uninitialized at line 459, and this was causing a Run-Time Check failure and crash in Debug versions on Windows.

I don't know whether 0 is the the correct value in this case, but this fixes the crash. mirandac@ will need to review the behavior.

BUG=None
TEST=Run debug version of Chrome on Windows and wait a few seconds.
Review URL: http://codereview.chromium.org/6385004