Reland 72784
This is the first step to being able to re-constitute policy from a non-generic protobuf in a policy request in a fully automated way (i.e. the code for the process can be generated). The non-generic protobuf is required for automatic merging of policy on the server-side.

BUG=68309
TEST=existing policy/provider tests

Review URL: http://codereview.chromium.org/6357021