Makes it so the hung renderer dialog isn't shown if the page that is hung is being shown by instant.
This also fixes a crash if the instant preview hung. The crash occurred because when HungRendererDialogView::ShowForTabContents set the bounds a synchronous focus change was triggered, which caused the instant preview to be destroyed, and when  hung_pages_table_model_->InitForTabContents(contents) was invoked the contents was destroyed and no renderer.

BUG=58853
TEST=see bug

Review URL: http://codereview.chromium.org/6286055