Update licenses.py to handle V8 and update about:credits.
The licenses.py changes haven't been approved for trunk yet, 
but the more important thing is to get about:credits up to date.

So, ignoring whether how we do it is well-coded, this makes sure
that V8 and Strongtalk licenses are properly credited in about:credits.

BUG= None
TEST= Find V8 and Strongtalk in about:credits.
Review URL: http://codereview.chromium.org/6286134