PhishingClassifierDelegate test cleanup, and fix a bug with session history navigation.
As of r74028, the RenderView will send notifications directly to the PhishingClassifierDelegate that we are using in the test.  As a result, we don't need to explicitly send DidCommitProvisionalLoad notifications to the delegate; the RenderView is doing that for us.

As a side-effect of replacing these one-off notifications with real navigations, I also noticed a problem where pending classifications are not cancelled on back/forward/reload because these are not new navigations.  This CL fixes that as well.

BUG=none
TEST=PhishingClassifierDelegateTest

Review URL: http://codereview.chromium.org/6425003