Revert 74887 - Add plumbing to webRequest API to support event filtering.
This introduces a new kind of event (WebRequestEvent) to handle the extra parameters to addListener. Behind the scenes, adding a listener creates a unique sub-event associated with that set of filter+extraInfo parameters. When we dispatch the event, we dispatch only those sub-events that match the required filters.

BUG=60101
TEST=no

Review URL: http://codereview.chromium.org/6250152

TBR=mpcomplete@chromium.org
Review URL: http://codereview.chromium.org/6526007