Merge 73251 - Fix a possible access to a deleted MetricsLog object.
In the case where StopRecording is called and the log is over the event limit, the current_log_ is deleted and a new one is created.  However, the local variable current_log was still pointing to the old object.  Fixed by moving the initialization of current_log to after the log is recreated.  This was introduced in http://codereview.chromium.org/2731006 .

BUG=64813
TEST=none

Review URL: http://codereview.chromium.org/6347034

TBR=bryner@chromium.org
Review URL: http://codereview.chromium.org/6480064