Merge 75043 - Simple scheduling for requestAnimationFrame
This simplifies the schedule for requestAnimationFrame callbacks and always invokes
the callbacks from the top of DoDeferredUpdate.  Also adds a histogram for catch-up
timers so we can more easily quantify the effects of base::Time / base::TimeTicks
skew.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6519005

Needed to ensure that the bugfix for 71735 merges cleanly to the branch.

TBR=jamesr@chromium.org
Review URL: http://codereview.chromium.org/6507021