Fix uninitialized start time in queued safe-browsing checks.
While loading the database, checks can be queued.  After loading the
database, the time-in-queue is histogrammed, but the start time was
never initialized, so the histogram could fail a DCHECK().

Also convert base::Time usage to base::TimeTicks, so that time never
flows backwards.

BUG=73072
TEST=none

Review URL: http://codereview.chromium.org/6528051