Fix is_convertible for windows.
It turns out that if you nest a template inside the initializer list of another template, the namelookup in windows gets all funky.  In this case, it couldn't tell that the found name was a function.

Luckily, we could just move the template parameters one level deeper.

BUG=none
TEST=unittests & trybots.

Review URL: http://codereview.chromium.org/6537001