Add a mode to archive build to support uploaded to google storage.
This currently looks pretty much like a hack, but once this land, and
everything is working, we will be able to remove the old mode to
archive to a server. This should simplify a lot the code, and remove
most of the hacks.

The current approach:
  Override all calls writing to the server, and replace them
  with a call to gsutil if it is needed.

Review URL: http://codereview.chromium.org/6524028