Intergration of the client-side phishing detection.
If the client-side phishing detection classifies a page as phishing it will
send back a ping to Google to verify whether or not the page is really phishing.

If the server also classifies the site as phishing we may show a phishing
interstitial if it is enabled.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6014003