Make chromoting connect always work
There was a race condition in the chromoting extension javascript. We wait
for OnUpdated() in order to send the chromoting tab a request for connection.
However OnUpdated() is insufficient for the chromoting tab is fully loaded.
And thus sending a request may not always work. This change will wait for the
tab to go to a "complete" status before a request is sent.

BUG=73202
TEST=Click on chromoting connect, it should always connect

Review URL: http://codereview.chromium.org/6532042