Callbacks: Replumb the type-inference.
In preparation for attempting to support __stdcall, __fastcall, etc., break
apart the templates so that instead of having one set of InvokerNs per type
of function pointer, only have one per syntactic method of invocation.

This lets the number of template specializations scale better.  Previously,
for each type of function pointer, we needed sum(1...arity) InvokerN
specializations.  There were 3 types (function, method, const method).
The Windows calling conventions would have added another 2.

in this method, we have 2 sets of InvokerN templates, and 1 set of
FunctionTraits templates for each type.  We only need (arity) number of
FunctionTraits templates, so this is a net win.  For our 6-arity system,
it should go from

  5 types * (1+2+3+4+5+6) InvokerNs  = 105 specializations

to

  5 types * 6 FunctionTraits + 2 calling_syntaxes * (1+2+3+4+5+6) InvokerNs = 72 specializations

This puts a bit more work on the compiler, but...hey, better it than the reader of the code.

BUG=35223
TEST=none

Review URL: http://codereview.chromium.org/6538045