Allow the reliability tests to run more than on URL in JS stress mode
There was i bug in the reliability tests where passing --stress-opt or --stress-deopt would cause the loading of the second URL from the list passed with the --list argument did not wotk. This was not an issur for the reliability test runner as it currently only loads one URL in each test when running in JS stress mode.

Also changed the JS stress message to be a control message.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/6538078