Autofill crasher Chrome:+Crash+Report+-+Stack+Signature:+WebCore::Node::renderBoxModelObject()
Attempt at fix for new crasher specific to 11.0.680.0.  The Autofill code on the
renderer side has not changed in a couple weeks.  Upon inspection of
WebCore::Node::renderBoxModelObject() it would seem that the |this| pointer is
NULL.  This change adds an isNull() check on the node in question, but this
doesn't address the source of the issue.  Will follow-up with QE to bisect and
isolate if possible.

BUG=73895
TEST=observe crash reports on canary

Review URL: http://codereview.chromium.org/6577009