Move some more direct uses of RootView to Widget.
The idea is to remove RootView from the public API of Views and eventually move it into the internal namespace. It is really an implementation detail of event propagation into a view hierarchy.

Anyone that calls GetRootView() and does something that isn't a View method should be calling that method on the Widget instead...

e.g. instead of calling GetRootView()->NotifyThemeChanged(), call GetWidget()->ThemeChanged().

The RootView is a focus traversable, which is fine, but calling code should call GetWidget()->GetFocusTraversable() to obtain it, rather than knowing RootView is-a focus traversable and just calling GetRootView().

This also changes FocusManager::ContainsView to be simpler and cross platform. Since there is one FocusManager per view hierarchy (attached to the toplevel Widget), getting the view's focus manager and comparing it to the current focus manager is a sufficient test.

BUG=72040
TEST=existing unittests, none
Review URL: http://codereview.chromium.org/6577017