Http cache: Fix handling of truncated entries.
- Always validate a truncated entry. Furthermore, given
that a server can say not-modified (304), followed  by
200 when actually asked for the final range, now we
validate truncated entries by asking for the next byte
that we need.

- If a server refuses to serve a byte range and instead
sends the whole resource, we just cache the response as
if nothing was stored before.

- If we receive a network error while reading from the
net, we don't discard the cached entry right away. Now
we wait until the transaction is closed, and attempt to
mark the entry as truncated.

BUG=48468
TEST=net_unittests

Review URL: http://codereview.chromium.org/6588105