Revert 76792 - fix display of filenames in file:/// URLs It breaks the following tests: BrowserTest.NoTitle and AutomationProxyTest2.GetTabTitle
The code attempted to shorten file:/// URLs to just the filename
when displayed as the title of a page.  But that appears to have
regressed sometime in the past.  This shortening is consistent with
how we display the title of images (which are like "foo.png (123x456)".)
Chrome does a poor job of displaying longer titles (most of the
tab title ends up being "file:///C:/" anyway).

In any case, using a FilePath to get the filename from a URL may not
have even worked on Windows, where the path separator is a backslash.

It appears Glen wrote the original code, Brett may be the one to have
regressed it in a refactor, and I probably broke it worse in a FilePath
refactor.

BUG=69467
TEST=load a text file via a file: URL; tab title is just the file name

Review URL: http://codereview.chromium.org/6591127

TBR=evan@chromium.org
Review URL: http://codereview.chromium.org/6614025