Relanding this.
Enable ChromeFrame net tests. These tests regressed with the change in the chunked upload
code in net to move the formatting of individual chunks to the http network stack.

This caused the URLRequestTestHTTP.TestPostChunkedDataBeforeStart test to hang in ChromeFrame.

Fix is to ensure that the individual chunks are properly formatted while marshaling them over
IPC.


BUG=none
TEST=chrome frame net tests should pass.
TBR=amit

Review URL: http://codereview.chromium.org/6621024