Add brand-code migration to the Chrome installer when creating the new multi-install app guid.
When upgrading a product from single to multi, the installer will now check if a Chrome install at that level (user/system) had a brand code, and if so it will copy the brand code to the ClientState key for the multi-install.

Also, some yak de-hairing: alter InstallationState such that it can be queried for partial ProductStates for products that are in the process of being installed. This can be used to query e.g. the ap value and the brand code, but not the version (since a product won't have a version until its first install is complete). This removes the need to go and manually re-query the registry in a couple of places, makes the installer's actions based solely on the original InstallationState and generally removes some hackiness.

BUG=61609
TEST=setup_unittests.exe, also when upgrading from single to multi, observe that brand codes are copied over.


Review URL: http://codereview.chromium.org/6604010