SUMMARY: Create dropdown menu button on Windows Aero frame to be displayed when the user is synced. This is the first step in the multi-profile UI implementation on Windows. The displayed tag consists of two parts:
1. A profile menu button with invisible border and background consisting of text and a dropdown arrow.
2. A tag image displayed on the browser frame beneath the profile menu button.


DETAILED CL DESCRIPTION:

1. Changes to the frame and view

* glass_browser_frame_view
adds the profile_menu button to the frame, along with methods for laying it out, and watching for user name changes.

2. New profile menu button

* profile_menu_button
sets up the actual menu button itself, and provides the method PaintProfileTag that draws the underlying tag image on the browser frame.  a better way to do this might be to create a new kind of menu button that combines methods from ImageButton as well as TextButton -- but maybe that's overkill?
Review URL: http://codereview.chromium.org/6490022