Add common.manifest to load a manifest and its dependencies.
It's made so it's easily extensible and clean and easy to write while supporting
multiple platform.

It's not exactly the same format as the one used by slavelastic, the 'other'
manifest format will need to be ported.

Next I'll add isolate that uses this code. Isolate is a tool to run a test
locally in a temporary directory to verify the input files.

R=nsylvain@chromium.org

Review URL: http://codereview.chromium.org/6683042