[Mac] Tighter restrictions on NSException raises.
On the crash server, almost all NSInvalidArgumentException cases are
due to sending a bad selector to an object.  Most likely cause is a
stale pointer which has been reallocated to another object, or memory
stompers.

Also start tracking uncaught NSInternalInconsistencyException, which
is the most popular exception on the crash servers.

BUG=24462
TEST=none

Review URL: http://codereview.chromium.org/6698002