Make URLRequestHttpJob the only URLRequestJob concerned with FilterContext.
In order to allow this, provide a new factory method in Filter to return a
GZipFilter (used by URLRequestFileJob and URLRequestJobTrackerTest), and
invert control so that URLRequestJobs are responsible for constructing filters.

This is one step away from removing FilterContext as a base class of URLRequestJob,
which will be tackled in a followup change.

BUG=none
TEST=net_unittests

Review URL: http://codereview.chromium.org/6674042