History tests de-flaked.
The history tests used to be flaky on all platforms.
Several fixes were made, and now the flakiness dashboard shows that the only source of failures is due to adding a URL and searching right away on windows XP.  The test that fails has been isolated in a single subtest (ExtensionHistoryApiTest.SearchAfterAdd), and all other tests are no longer marked as flaky.

BUG=26296
TEST=ExtensionHistoryApiTest.*

Review URL: http://codereview.chromium.org/6698028