Normalize the top-level ui/ module.
This CL creates four .gypi files that are included in /ui/ui.gyp. This allows
for a single unittest binary for all the submodules, rather than having
individual unittest binaries for each submodule.

To not break the buildbot, this creates a fake target that copies ui_unittests
to gfx_unittests as temporary scaffolding.

BUG=72317
TEST=gfx_unittests (really ui_unittests) passes

Review URL: http://codereview.chromium.org/6688007