Do not put off destroying of the IOSurface in the GPU process.
Evicting old graphics contexts will be done in the GPU process by removing
routes from GpuChannel::router_.  Doing so on a mac results in an assert when
handling GpuCommandBufferMsg_SetWindowSize.  The assert was a consequence of how
we deal with closed windows.  When a window is closed, the browser process sends
a message to the GPU process to request that the surface is removed.  Rather
than doing that immediately, GpuThread object stores the view id of the renderer
in destroyed_renderer_routes_.  GpuCommandBufferStub::OnFlush always checks if
its view id is in that set, and if is, it asks the GPUProcessor to destroy its
surface.

This change gets rid of the destroyed_renderer_routes_.  Rather than deferring
the request to destroy the surface, we ask the GPU process to remove the route
which, in turn, will remove the GpuCommandBufferStub, GPUProcessor and its
surface.

See also bug 67170.

BUG=none
TEST=set the breakpoint in GpuChannel::OnMessageReceived, manually remove the route from router_ and stubs_.  No assert should be triggered and context should be silently recovered.

Review URL: http://codereview.chromium.org/6688026