Provisional fix for WebFrame API change upstream (webkit bug 56139).
This is somewhat hackish.. the new version of the function upstream has a default initializer. I think it's better though to add it this way to remove ambiguity - for the overloading the default initializer should not matter.

Either way, I expect to remove the obsolete definition within a day or so.

Review URL: http://codereview.chromium.org/6677069