Inform user to reload page after chagning cookie settings from in-page dialog.
NOTE: the png file is not final yet, waiting for chrome ui folks to provide
final vesion.  However, this does not change the code, so please review.
Thanks.

BUG=63645
TEST=Block all cookies and load a page that uses cookies.  Click on the broken
cookie icon in the omnibox and open the in-page cookie dialog.  Change a
blocked to allow, or block an allowed cookie.  After clicking the Close button
an infobar appears with a warning.  Make sure the reload button works.
Review URL: http://codereview.chromium.org/6676041