Added a new states to the CloudPolicyCotroller - STATE_TOKEN_ERROR, STATE_TOKEN_UNMANAGED.
This STATE_TOKEN_ERROR is entered if the DMServer suffers some form of amnesia and tells us that the device token is unknown just after it gave it to us. In this case we will refetch the token but instead of doing this full steam
we will be gracefully backing off with time. Also whenever we get a response that a policy is not present for a device we will accept this an OK state and try to refetch only after the regular policy update interval.

BUG=76507
TEST=Change the testserver response of policy fetch request to DEVICE_MANAGEMENT_NOT_SUPPORTED and observe that the server is not DoSed with policy requests.

Review URL: http://codereview.chromium.org/6670088