Merge r79013 -- Direct Windows users to a better Java download page.
There are a couple of cases where we either redirect to a generic Java info page (http://java.com/) or
to a locale-specific page.

http://java.com/download does OS and locale detection automatically, so we should redirect users there.

The page does not handle linux specially, so I left the URL alone for Linux in plugin_list.cc.

BUG= http://bugs.chromium.org/76634

TEST= On a Windows machine without Java installed or with an out of date version of Java, going to a site that uses the Java plugin
should show the missing plugin link and link to http://java.com/download... never http://java.com/

R=jam@chromium.org, darin@chromium.org, ananta@chromium.org, panayiotis@google.com