Die, DemuxerStream::QueryInterface, die!
This appears to have been an attempt at hand-rolling an RTTI system, all for the
use of a *single* class (!).  Removing it makes for cleaner, simpler, shorter
code.  It also enables delegating DemuxerStreams, which wasn't possible before
(b/c the overridable method was protected, so not visible to a HAS-A wrapper).

BUG=none
TEST=trybots

Review URL: http://codereview.chromium.org/6730006