Notify child widgets before and after a reparenting.
Attaching a native view to a NativeViewHostWin may cause that
view to be disassociated from a view hierarchy with a different
FocusManager.  Notifying the widgets that they are about to be
removed from their current parent gives them the opportunity to
disassociate themselves from their current FocusManager.

BUG=77447
TEST=none


Review URL: http://codereview.chromium.org/6776025