GTTF: Detect browser crashes on shutdown in UI tests.
Previously the automation framework could miss a browser
crash during shutdown on POSIX (on Windows there is
crash_service.exe that should catch all crashes).

This change makes the automation framework avoid losing
information about the browser process' exit status
(CrashAwareSleep), and fixes a bug in base::WaitForExitCodeWithTimeout
(which on POSIX never reported the process has been signaled).

Finally, it makes the automation framework use WaitForExitCodeWithTimeout
instead of WaitForSingleProcess. This way we can get the exit status
information in an accurate and cross-platform way.

To avoid trying to close the same process handle twice (it's only an issue on Windows) I've changed WaitForExitCodeWithTimeout not to close the passed handle. It's only used in few places and I think this CL fixes all of them.

I've tested this change locally on Mac with a UI test that SIGKILLs the browser.
Before this change the test passed (it shouldn't), and after this change
the test failed with an information that the browser has not exited cleanly.

BUG=56644
Review URL: http://codereview.chromium.org/6689014