Adds two Python/PyAuto functions, GetProxySettings() and SetProxySetting(), that allow tests to read and manipulate Chrome OS proxy settings.
Fixes issue with proxy settings where they reset when changing only the port of a proxy without changing the URL. This happens when using the automation hooks because they create a new ProxyCrosSettingsProvider with each invocation. This is done by removing the caching behavior of ProxyCrosSettingsProvider.

Modifies the type of the "Port" textboxes in the Proxy Settings UI to Integer so that ProxyCrosSettingsProvider uses a consistent type for both Get() and Set().

BUG=chromium-os:13650,chromium-os:12727
TEST=Manual. Test included for GetProxySettings() that will run in the Autotest lab. stanleyw will add additional tests that use SetProxySetting().

Review URL: http://codereview.chromium.org/6737035