[Mac] Don't route mouse events to the render view host if window is not frontmost.
This change makes Mac Chrome match the behaviour of other browsers on the Mac
(tested Firefox and Safari) as well as other Mac applications such as iTunes.

Additionally, this fixes a number of bugs, including:
- Mouse cursor changing when hovering over web content in background windows
yet clicks in this case would be eaten to bring the window to the front.
- Mouse cursor NOT changing when hovering web content in background windows when
Chrome is not the active application - resulting in inconsistent behaviour.
- http://crbug.com/77928 - resulting in wrong cursor persisting after moving out
of the web content area of a background window.

Also, change related unit tests to make the test window the keyWindow so that the
tests continue to pass.

BUG=77928
TEST=Open two windows, A and B. Open a webpage in A, then make B the foreground
window. Now, hover the cursor over web content in A. The cursor should not
change and mouse-over actions should not occur.
R=mark@chromium.org

Review URL: http://codereview.chromium.org/6720023