Don't discard a header parameter because of unbalanced quotes.
Some web servers reportedly send header parameters that aren't
quoted properly.  So we end up with headers like:

  Content-Disposition: attachment; filename="Foo.bar

Rather than discarding the filename parameter in this case, we
now treat it as if there was a closing quote at the end of the
line.

BUG=58840
TEST=net_unittests --gtest_filter=NetUtilTest.GetFileNameFromCD


Review URL: http://codereview.chromium.org/6802019