Detect download errors.
When a download error occurs, we now mark it as "Interrupted", and display the new status on the Download Shelf.

This is handled through the OnResponseCompleted() method, from:
DownloadResourceHandler ->
  DownloadFileManager ->
    DownloadManager

The latter checks the error value, and if not zero calls DownloadError().

DownloadError() marks the download as interrupted (stopping the progress timer), updates the download history, and updates its Observers.

DownloadItem::DownloadState now has a new state, INTERRUPTED.

There are also new display strings that show up on interrupted downloads in the download shelf and download page.

BUG=35278
TEST=Pass unit and browser tests.  Also, run a python server that behaves as described below, connect to it, and observe that "Interrupted" shows up in the item on the download shelf.  Check that the item behaves as if it were cancelled.

The Python server should respond to a GET request with a content length that is large, then provide several chunks of data (but less than the specified length) and then send a 'No file' response (code 204) and close the connection.

Review URL: http://codereview.chromium.org/6613024