Ensure User-agent is present in preconnection request
Use the global default context to get *some* UA
string in a preconnect.  This is a change that was
suggested by wtc, and I've tested it by adding
(in my local build) a CHECK() of the getter
result in net/socket/client_socket_pool_manager.cc.

This change is a very low risk, high reward P1 fix.
It is pretty inconceivable that it could do anything
other than help the status quo.

I was going to land the CHECK(), but the number of
changes to unit tests was large, and adding the
CHECK() is best done on a canary.  As a result, I'll
break out the changes to at least:

      net/http/http_stream_factory_impl_unittest.cc
      net/http/http_network_transaction_unittest.cc

along with the CHECK() change to:

      net/socket/client_socket_pool_manager.cc

in a separate CL.


r=wtc
BUG=78037
Review URL: http://codereview.chromium.org/6799018