Send the ref as part of the client-side phishing detection pingback.
This will help determine whether differing scores on the same URL are related
to data that is passed through the ref (as in Google Instant queries).  Also
add a test for the url feature extractor to verify that the ref is not used to
form any token features.

BUG=none
TEST=PhishingClassifierDelegateTest,PhishingUrlFeatureExtractorTest

Review URL: http://codereview.chromium.org/6812009