Remove unneeded v8 lockers from idb_bindings.cc.
Since this code is pulled into webkit_unit_tests, it is making us
need lockers there too, which stinks.

I don't see any reason for these to exist. My speculation is that
this code used to run in the browser process and was needed at
that time.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6812038