Add parser and test files
This CL adds the PLY based parser, parser test files, and makes some additional
cleanup to the lexer.  The parser is part of a generator which is still under
development.  The parser can be run standalone which works as a test mode
verifying success and failure tags it find in the sources embedded as 
comments.  Adding the switch  --output will print the AST.


R=sehr@google.com
BUG=77551
TEST=python idl_parser.py test_parser/*.idl
Review URL: http://codereview.chromium.org/6754001