Fix crash in fade tab title change
The fade tab title change is causing the browser to crash on some computers.

It looks like the problem is in BitmapPlatformDevice::create(). For some reason it's returning NULL.

This change just adds a NULL check and falls back to the old drawing code.

Once this is checked in I'll investigate what the root cause is.


BUG=78857
TEST=Compiled and ran.

Review URL: http://codereview.chromium.org/6823020