Minor html fix and changed the size of the control links in interim login screen
I have been having issues with clicking the targets on this page during development do to other issues. Since this is not intended to be the final screen I just increased the size of the text so I don't accidentally click the wrong link. This caused it to be obvious that one of my divs wasn't closed correctly, so I fixed that also.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual


Review URL: http://codereview.chromium.org/6801048