Change the V8 buildbot master to support 3.* branches and trunk.
This change adds support for svn pulling the 3.0, 3.1, and 3.2
branches and the trunk repository in addition to the current pull of
bleeding edge. In addition, it creates 4 new slaves (that will run of
the same machine, it would be vastefull to use 4 machines for this
since this will be something around 10 runs per week on average - 2
pushes to trunk plus some patches).

I also changed the name of the build directory to better reflect that
this is not necessarily bleeding edge (people would probably find it
odd to see a bleeding_edge path on a buildbot running on trunk).
Review URL: http://codereview.chromium.org/6849026