Remove unused code from MetricsService
Per discussion, this code is essentially all unused (the server never sends the node this looks for at the top level), and we don't expect to need it at this point, so take it out. If we do need some of this capability, we can add pieces back as necessary--with tests!

BUG=None
TEST=Existing MetricsServer tests, but mostly watching for unexpected changes in dashboards.

Review URL: http://codereview.chromium.org/6862001