[Sync] Fix race condition in P2PNotifier with sending notifications
Store any notifications sent when not connected and send them on
the next connect.

This fixes a race condition in the sync integration tests exposed
by the new syncer thread.

Make sending of notifications not be blocked on successful subscription.

Disable a failing test in jingle_unittests.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6881042