Add 'Cache-Control: no-cache' to filesystem: responses to avoid having them improperly cached in WebKit's memory cache.
Note that this change as currently written may be unnecessarily
bad for performance, since the file must be read from disk on
every access. For some use cases, this could be quite expensive.

The right fix is probably to put a 'Last-Modified' header in the
response along with 'Cache-Control: max-age: 0', but for now
I'm favoring correctness over performance.

R=michaeln@chromium.org
BUG=79539
TEST=layout test at http://webkit.org/b/58854

Review URL: http://codereview.chromium.org/6879022