Rename NaClThread to NaClLauncherThread
This avoids a conflict with NaCl's own "struct NaClThread".
NaClLauncherThread is also a more descriptive name.

The goal is to be able to #include NaCl's sel_ldr.h so that we do not
have to duplicate the prototype for NaClMainForChromium().  However,
we cannot do that immediately because there is still a conflict
between NaCl's LOG_FATAL and Chromium's LOG_FATAL.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1595
TEST=trybots

Review URL: http://codereview.chromium.org/6880074