[Mac] Show poof animation for current window only when deleting BMB bookmarks
Checks |isKeyWindow| before performing the poof animation. Re-ordered the
call to [oldButton removeFromSuperview] so that the view will actually have
a window at the time of the check. Also, moved initialization of |poofPoint|
inside the block to avoid doing the work when it's not needed.

BUG=80046
TEST=See bug.

Review URL: http://codereview.chromium.org/6883121