GTK implementation of "Select tabs when dragging onto them".
This is the GTK-side of:

http://codereview.chromium.org/6806013/

Which implements this feature on Windows and Mac. (This CL depends on that CL, which I've not landed yet.)

BUG=19071
TEST=1. Select some text and drag it over an inactive tab. In 0.5 seconds,
the tab selection should switch to that tab. You should now be able to
drag the text into the content area of that tab (e.g. into a text box).
2. Select some text and drag it over an inactive tab. Without waiting
over 0.5 seconds, move the mouse away from that tab. No tab switch
should take place.

Review URL: http://codereview.chromium.org/6869012