Fix some unit test failures.
These failures only happen if tests are called in the right (wrong?) order.

ExtensionSourcePriorityTest::SetUp() failied to call SetUp on it's parent class.  If a previous test caused an error, it would fail with that error.

ExtensionServiceTestSimple.Enabledness should initialize ExtensionErrorReporter.  If it was run before other tests had done so, it would CHECK().

BUG=79755
TEST=ExtensionSourcePriorityTest.*, ExtensionServiceTestSimple.*


Review URL: http://codereview.chromium.org/6877072