Reland 81998.The bug in webkit\gpu\webgraphicscontext3d_in_process_impl.cc that assumed creating a new GLContext made it current is now fixed and this patch should now work.
Original message:

Split OSMesa implementations of *GLContext into GLContextOSMesa and *GLSurfaceOSMesa.

Surfaces are independent of contexts in GL. To facilitate sharing of surfaces between processes, I have separated them from the notion of contexts because contexts cannot be shared between processes.

I started with EGL in r81512 and WGL in r81807. This is the same thing for OSMesa.

GLContextOSMesa still has a pointer to a surface and still has some surface specific operations that just forward through to it. Once I have refactored all the GLContext implementations in this way, I will remove these pointers and the surface specific opertations.

There will not be "view" and "offscreen" GL contexts. Rather there will be a single context type for each backend which can be made current with a surface that directs output either to a view or offscreen surface.

Original review:
http://codereview.chromium.org/6864031/

TEST=ran linux_layout and mac_layout try jobs (they were failing before)
BUG=none
Review URL: http://codereview.chromium.org/6881071