Move the synchronous GPU messages to the IO thread to avoid deadlock.
I patched in Jonathan's change from http://codereview.chromium.org/6881105/ for the CreateViewCommandBuffer message, and also added the Synchronize and EstablishChannel.  The latter required making GpuDataManager callable from the IO thread.  I moved the code that loads the blacklist from the prefs and web to Chrome code, since I wanted to do that anyways so that GpuProcessHostUIShim and GpuDataManager can move to content (I'll do that later).

Since the messages are filtered on the IO thread, it's now GpuProcessHost that creates GpuProcessHostUIShim.  Accordingly, all the code that used to call GpuProcessHostUIShim to send a message now has to call GpuProcessHost, since the logic of when to create a process is there.  Also, since there's no IO thread object for the in-process case, I've had to break that in the meantime.  Al will take a look at that later.

BUG=77536 
Review URL: http://codereview.chromium.org/6902021