Allow gpu ringbuffer to wrap around.
The ringbuffer was unwilling to wrap around unless the last block filled it up completely - otherwise it would have to wait until it was completely empty. Fix this in the most straightforward way by filling the ringbuffer up with a padding block if necessary.

BUG=80946
TEST=

Review URL: http://codereview.chromium.org/6901138