[Mac] Fix dragging files from the download manager.
This stopped working when TabContentsViewCocoa was changed to support drags from a WebDragSource. That change made |draggingSourceOperationMaskForLocal:| just forward to the same method on the dragSource_, which the code assumed would always be set via |startDragWithDropData:|. In the case of drags from the download manager, which are initiated by DragDownload(), that is not the case.

This change makes TabContentsViewCocoa's |draggingSourceOperationMaskForLocal:| return NSDragOperationCopy if there is no |dragSource_| and adds a test that this is the case when |startDragWithDropData:| hasn't been called.

BUG=80833
TEST=Download a file. Open the downloads manager. Drag said file from the downloads manager to the desktop. The drop should take place and the file should be copied to the desktop.
Review URL: http://codereview.chromium.org/6927010