Ensure that sessions encryption sync tests wait for OnPassphraseRequired
The sessions encryption sync integration tests are flaky on the sync
buildbots because they assume that a passphrase is required on Client 1
as soon as Client 0 sets a custom passphrase and a sync cycle completes.

This patch adds a method called AwaitPassphraseRequired to
ProfileSyncServiceHarness, so that Client 1 can first wait until it
realizes that a passphrase is required before calling SetPassphrase. It
also fixes the integration tests that were flaky.

BUG=81688
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6949003