Merge 85061 - Allow the first-run bubble to be shown on organic first runs.
The code had some mysterious chokepoint that prevented the bubble from being shown if the attached parent window wasn't active. However the first run bubble is shown from the browser window's construction/init pass, before it's ever shown, so it can't be active.

It turns out the chokepoint was to prevent the bubble from being shown later as the TemplateURLModel is loaded, so I moved the active check to that part of the code so that the initial display could work.

http://crbug.com/78878
TEST=run chrome with --user-data-dir=c:\temp\test --first-run --organic (the user-data-dir MUST be empty!) - first run bubble should appear. also, browser window should render as active for at least 5-10 seconds after displaying. this code fixes a bug where running with just --first-run would have the browser appear inactive after about 3 seconds.
Review URL: http://codereview.chromium.org/6973012

TBR=ben@chromium.org
Review URL: http://codereview.chromium.org/7015011