Revert 85269 - Split the hierarchy.
* Widget ---- 

Now recognizes a supplied NativeWidget via InitParams. If this is specified then a default one is not created. 

Is now created directly rather than using a factory. NativeWidget creation is not performed until Init() is called. This means some functions that rely on a NativeWidget must not be called until _AFTER_ Init() (explains some of the function call reordering in this CL, e.g. moving SetOpacity() until after Init()). 

ResetLastMouseMovedFlag() moved to this API so that BaseTabStrip can call it in a cross-platform way. 

Made last remaining unimplemented methods on Widget pass-thru to NativeWidget implementations. 

* WidgetWin/WidgetGtk ---- 

The NativeWidget implementations now both require a NativeWidgetDelegate implementation upon construction. This is passed through the constructor by the static factory method NativeWidget::CreateNativeWidget and by subclasses such as WindowWin, BubbleWidgetWin, etc. 

Some classes that are constructed directly (e.g. LockWindow, in ChromeOS) never have a Widget created for them, so they create the Widget themselves in their base class initializer. 

Code in these classes (and their WindowWin/WindowGtk, BrowserFrameWin, BrowserFrameGtk subclasses) must now call GetWidget() etc to call Widget API methods since they are no longer subclasses. 

static_casting to this (and derived) types must now be done on the Widget's native_widget(). 

GetWindow() is renamed to GetContainingWindow() to avoid naming conflicts. 

* Window ---- 

Window is now a subclass of Widget. 

Now recognizes a supplied NativeWindow via InitParams. If this is specified then a default one is not created. 

Window::CloseWindow becomes an override of Widget::Close. 

CloseAllSecondaryWindows() becomes CloseAllSecondaryWidgets() and moves to widget.h 
IsAppWindow() is removed and replaced by set_is_secondary_widget on Widget. 

* MenuHost ---- 

Subclasses Widget now. 

* TabContentsViewViews ---- 

It looks like the Gtk-views code here was still using the old implementation of the Native version of this class - i.e. a class that subclassed TabContentsView AND WidgetGtk. A no-no. I had to write NativeTabContentsViewGtk, which is almost identical to NativeTabContentsViewWin with the Gtk bits of TabContentsViewGtk thrown in. 

* BrowserFrame ---- 

Platform-specific functionality is now restricted to BrowserFrameWin/BrowserFrameGtk behind a NativeBrowserFrame interface. Construction is exposed via a static factory method on NativeBrowserFrame. 

BrowserFrame becomes a concrete class that now subclasses Window. 

As a result, it no longer needs a GetWindow() accessor method, so people with a BrowserFrame* can just call Window methods directly on it. 

It is constructed directly, replacing the BrowserFrame::Create() method. 

NativeBrowserFrameDelegate is no longer needed. 

BrowserFrameChromeos is simpler as a couple of #ifdefs in BrowserFrame, so I got rid of that too. 

* AutocompletePopupWin/Gtk ---- 

No longer required. AutocompletePopupContentsView now just uses a Widget directly. 

* There is some lingering ugliness: 

- If you set a native_window field on Window::InitParams you must also manually set widget_init_params.native_widget. I will make InitParams do more of this automatically later. 
- It'd be nice for the ContentsView to be specified via InitParams. I'll get to this later. 
- NativeBrowserFrame could probably disappear as an interface. It only exists to provide a couple of methods that may be implemented in other ways. 
- delete_on_destroy should now be an ownership directionality enum. I will do this later. 
- Secondary-widgetness should somehow be inferred from transience. Later. 
- set_focus_on_creation for both the NativeWidgets should probably move to Widget if it is really needed. 
- WidgetWin/Gtk::SetInitialFocus seems like it could move to Widget. 
- I need to clean up function order in some cases. 

BUG=72040
TEST=none
Review URL: http://codereview.chromium.org/7012006

TBR=ben@chromium.org
Review URL: http://codereview.chromium.org/7011038