Added release build assert on attempt to create a RunnableFunction for a function pointer with address 1.
This is actually happening. See http://crbug.com/81449. The generated code to invoke the callback puts the address of the function in the EAX register before doing CALL EAX. I see 0x00000001 in the EAX register when it crashes in the reported minidumps.

I'll revert this after the next Canary.

TEST=run locally and verify no assertion
BUG=81449
Review URL: http://codereview.chromium.org/7013014