MAC Cookies (patch 5 of N)
To help folks deal with clock skew, we've changed the timestamp to the "age" of
the credential, which is the number of seconds since the cookie was set (as
opposed to the number of seconds since the epoch).  This patch updates our
implementation to match.

Review URL: http://codereview.chromium.org/6969075